+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 ### 5.1.12
+
++ [x] Fixed method `val()` in `select` component to add value with icon